flask-microservices/auth-service/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Werkzeug==2.2.3
SQLAlchemy==2.0.36
PyJWT==2.8.0
